+++ .gitignore
@@ -35,6 +35,3 @@ out/
 
 ### VS Code ###
 .vscode/
-
-### yml ###
-/src/main/resources/application.yml
