+++ src/main/java/com/example/ordersystem/Controller/MenuController.java
@@ -1,7 +1,10 @@
-package com.example.ordersystem.order.controller;
+package com.example.ordersystem.Controller;
 
+import com.example.ordersystem.Service.OrderService;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
 
 @Controller
 public class MenuController {
