+++ src/main/java/com/example/ordersystem/DTO/OrderCreateRequest.java
@@ -1,0 +1,4 @@
+package com.example.ordersystem.DTO;
+
+public record OrderCreateRequest(String menuName, Integer quantity) {
+}
