+++ src/main/java/com/example/ordersystem/DTO/OrderGetResponse.java
@@ -1,0 +1,8 @@
+package com.example.ordersystem.DTO;
+
+import com.example.ordersystem.Entity.Order;
+
+import java.util.List;
+
+public record OrderGetResponse(List<Order> orders) {
+}
