+++ src/main/java/com/example/ordersystem/DTO/OrderUpdateRequest.java
@@ -1,0 +1,4 @@
+package com.example.ordersystem.DTO;
+
+public record OrderUpdateRequest(Long id, String menuName, Integer quantity) {
+}
