+++ src/main/java/com/example/ordersystem/Repository/OrderRepository.java
@@ -1,9 +1,12 @@
-package com.example.ordersystem.order.repository;
+package com.example.ordersystem.Repository;
 
-import com.example.ordersystem.order.entity.Order;
+import com.example.ordersystem.Entity.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
 public interface OrderRepository extends JpaRepository<Order, Long> {
+    Optional<Order> findById(Long id);
 }
