--- src/main/java/com/example/ordersystem/menu/entity/Menu.java
@@ -1,19 +1,0 @@
-package com.example.ordersystem.menu.entity;
-
-import jakarta.persistence.*;
-
-@Entity
-@Table(name = "menus")
-public class Menu {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long menuId;
-
-    @Column(unique = true, nullable = false)
-    private String menuName;
-
-    @Column(unique = true, nullable = false)
-    private Integer price;
-
-}
