--- src/main/java/com/example/ordersystem/order/dto/OrderCreateRequest.java
@@ -1,4 +1,0 @@
-package com.example.ordersystem.order.dto;
-
-public record OrderCreateRequest(Long menuFK, Long ordererFK, Integer quantity) {
-}
