--- src/main/java/com/example/ordersystem/order/dto/OrderGetResponse.java
@@ -1,8 +1,0 @@
-package com.example.ordersystem.order.dto;
-
-import com.example.ordersystem.order.entity.Order;
-
-import java.util.List;
-
-public record OrderGetResponse(List<Order> orders) {
-}
