--- src/main/java/com/example/ordersystem/order/dto/OrderUpdateRequest.java
@@ -1,4 +1,0 @@
-package com.example.ordersystem.order.dto;
-
-public record OrderUpdateRequest(Long id, Long menuFK, Integer quantity) {
-}
