--- src/main/java/com/example/ordersystem/orderer/dto/OrdererCreateRequest.java
@@ -1,4 +1,0 @@
-package com.example.ordersystem.orderer.dto;
-
-public record OrdererCreateRequest(String ordererName) {
-}
