--- src/main/java/com/example/ordersystem/orderer/repository/OrdererRepository.java
@@ -1,9 +1,0 @@
-package com.example.ordersystem.orderer.repository;
-
-import com.example.ordersystem.orderer.entity.Orderer;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface OrdererRepository extends JpaRepository<Orderer, Long> {
-}
