+++ src/test/java/com/example/ordersystem/OrderSystemApplicationTests.java
@@ -1,6 +1,12 @@
 package com.example.ordersystem;
 
+import com.example.ordersystem.Entity.Order;
+import com.example.ordersystem.Repository.OrderRepository;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+
+import java.util.List;
 
 @SpringBootTest
 class OrderSystemApplicationTests {
